Renaming challenge files to simplify and stop duplication in naming
