+++ packages/lib/src/components/ThreeDS2/components/Challenge/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ThreeDS2Challenge';
+export { default } from './PrepareChallenge3DS2';
