+++ packages/lib/src/components/ThreeDS2/components/Challenge/types.ts
@@ -2,21 +2,21 @@ import { ChallengeData, ChallengeObject } from '../../types';
 import { ResolveData } from '../utils';
 import { ThreeDS2ChallengeElementProps } from '../../ThreeDS2Challenge';
 
-export interface Do3DS2ChallengeProps extends ChallengeData {
+export interface DoChallenge3DS2Props extends ChallengeData {
     onCompleteChallenge: (resolveObject: ChallengeObject) => void;
     onErrorChallenge: (rejectObject: ChallengeObject) => void;
 }
 
-export interface Do3DS2ChallengeState {
+export interface DoChallenge3DS2State {
     base64URLencodedData?: string;
     status?: string;
 }
 
-export interface ThreeDS2ChallengeProps extends ThreeDS2ChallengeElementProps {
+export interface PrepareChallenge3DS2Props extends ThreeDS2ChallengeElementProps {
     onComplete?: (data: ResolveData) => void;
 }
 
-export interface ThreeDS2ChallengeState {
+export interface PrepareChallenge3DS2State {
     challengeData?: ChallengeData;
     status?: string;
 }
